refactor: revert serverpod_cloud_storage package
