+++ integrations/serverpod_cloud_storage_gcp/lib/src/native/native_google_cloud_storage.dart
@@ -9,7 +9,7 @@ import 'package:mime/mime.dart';
 import 'package:path/path.dart' as p;
 import 'package:pointycastle/asn1.dart';
 import 'package:pointycastle/export.dart';
-import 'package:serverpod_cloud_storage/serverpod_cloud_storage.dart';
+import 'package:serverpod/serverpod.dart';
 
 /// GCP cloud storage using native Google Cloud JSON API.
 ///
@@ -23,10 +23,7 @@ import 'package:serverpod_cloud_storage/serverpod_cloud_storage.dart';
 /// Example:
 /// ```dart
 /// pod.addCloudStorage(await NativeGoogleCloudStorage.create(
-///   passwordProvider: CloudStoragePasswordProvider(
-///     getPassword: pod.getPassword,
-///     loadPasswords: pod.loadCustomPasswords,
-///   ),
+///   serverpod: pod,
 ///   storageId: 'public',
 ///   bucket: 'my-bucket',
 ///   public: true,
@@ -60,21 +57,19 @@ class NativeGoogleCloudStorage extends CloudStorage {
        _authClient = authClient,
        super(storageId);
 
-  /// Creates a new [NativeGoogleCloudStorage] from a [passwordProvider].
+  /// Creates a new [NativeGoogleCloudStorage] from a [Serverpod] instance.
   ///
   /// The service account JSON is loaded from the password system via:
   /// - `passwords.yaml`: `gcpServiceAccount` key with JSON content
   /// - Environment variable: `SERVERPOD_PASSWORD_gcpServiceAccount`
   static Future<NativeGoogleCloudStorage> create({
-    required CloudStoragePasswordProvider passwordProvider,
+    required Serverpod serverpod,
     required String storageId,
     required String bucket,
     required bool public,
     String? publicHost,
   }) async {
-    final serviceAccountJson = passwordProvider.getPassword(
-      'gcpServiceAccount',
-    );
+    final serviceAccountJson = serverpod.getPassword('gcpServiceAccount');
     if (serviceAccountJson == null) {
       throw StateError(
         'GCP service account not configured. '
@@ -96,8 +91,8 @@ class NativeGoogleCloudStorage extends CloudStorage {
   /// Creates a [NativeGoogleCloudStorage] from a service account JSON string.
   ///
   /// This factory authenticates directly with the provided JSON credentials
-  /// without requiring a [CloudStoragePasswordProvider], making it suitable
-  /// for integration testing and standalone usage.
+  /// without requiring a [Serverpod] instance, making it suitable for
+  /// integration testing and standalone usage.
   static Future<NativeGoogleCloudStorage> fromServiceAccountJson({
     required String storageId,
     required String bucket,
@@ -249,7 +244,7 @@ class NativeGoogleCloudStorage extends CloudStorage {
 
   @override
   Future<void> storeFile({
-    required CloudStorageSession session,
+    required Session session,
     required String path,
     required ByteData byteData,
     StoreFileOptions options = const StoreFileOptions(),
@@ -296,7 +291,7 @@ class NativeGoogleCloudStorage extends CloudStorage {
 
   @override
   Future<ByteData> retrieveFile({
-    required CloudStorageSession session,
+    required Session session,
     required String path,
   }) async {
     try {
@@ -327,7 +322,7 @@ class NativeGoogleCloudStorage extends CloudStorage {
 
   @override
   Future<FileStat> statFile({
-    required CloudStorageSession session,
+    required Session session,
     required String path,
   }) async {
     try {
@@ -366,7 +361,7 @@ class NativeGoogleCloudStorage extends CloudStorage {
 
   @override
   Future<Uri> publicDownloadUrl({
-    required CloudStorageSession session,
+    required Session session,
     required String path,
   }) async {
     if (!public) {
@@ -382,7 +377,7 @@ class NativeGoogleCloudStorage extends CloudStorage {
 
   @override
   Future<Uri> temporaryDownloadUrl({
-    required CloudStorageSession session,
+    required Session session,
     required String path,
     TemporaryDownloadUrlOptions options = const TemporaryDownloadUrlOptions(),
   }) async {
@@ -423,7 +418,7 @@ class NativeGoogleCloudStorage extends CloudStorage {
 
   @override
   Future<void> deleteFile({
-    required CloudStorageSession session,
+    required Session session,
     required String path,
   }) async {
     try {
@@ -436,7 +431,7 @@ class NativeGoogleCloudStorage extends CloudStorage {
 
   @override
   Future<UploadDescription> createUploadDescription({
-    required CloudStorageSession session,
+    required Session session,
     required String path,
     UploadOptions options = const UploadOptions(),
   }) async {
@@ -603,7 +598,7 @@ class NativeGoogleCloudStorage extends CloudStorage {
 
   @override
   Future<bool> verifyUpload({
-    required CloudStorageSession session,
+    required Session session,
     required String path,
   }) async {
     return fileExists(session: session, path: path);
