+++ integrations/serverpod_cloud_storage_s3_compat/lib/src/client/exceptions.dart
@@ -1,5 +1,5 @@
 import 'package:http/http.dart' as http;
-import 'package:serverpod_cloud_storage/serverpod_cloud_storage.dart';
+import 'package:serverpod/serverpod.dart';
 
 /// Exception thrown when an S3-compatible API returns an error.
 class S3Exception extends CloudStorageException {
