+++ integrations/serverpod_cloud_storage_s3_compat/lib/src/s3_compat_cloud_storage.dart
@@ -2,7 +2,7 @@ import 'dart:io' show HttpDate;
 import 'dart:typed_data';
 
 import 'package:http/http.dart' as http;
-import 'package:serverpod_cloud_storage/serverpod_cloud_storage.dart';
+import 'package:serverpod/serverpod.dart';
 
 import 'client/exceptions.dart';
 import 'client/s3_client.dart';
@@ -98,7 +98,7 @@ class S3CompatCloudStorage extends CloudStorage {
 
   @override
   Future<void> storeFile({
-    required CloudStorageSession session,
+    required Session session,
     required String path,
     required ByteData byteData,
     StoreFileOptions options = const StoreFileOptions(),
@@ -137,7 +137,7 @@ class S3CompatCloudStorage extends CloudStorage {
 
   @override
   Future<ByteData> retrieveFile({
-    required CloudStorageSession session,
+    required Session session,
     required String path,
   }) async {
     final response = await _client.getObject(path);
@@ -155,7 +155,7 @@ class S3CompatCloudStorage extends CloudStorage {
 
   @override
   Future<FileStat> statFile({
-    required CloudStorageSession session,
+    required Session session,
     required String path,
   }) async {
     final response = await _client.headObject(path);
@@ -206,7 +206,7 @@ class S3CompatCloudStorage extends CloudStorage {
 
   @override
   Future<Uri> publicDownloadUrl({
-    required CloudStorageSession session,
+    required Session session,
     required String path,
   }) async {
     if (!public) {
@@ -221,7 +221,7 @@ class S3CompatCloudStorage extends CloudStorage {
 
   @override
   Future<Uri> temporaryDownloadUrl({
-    required CloudStorageSession session,
+    required Session session,
     required String path,
     TemporaryDownloadUrlOptions options = const TemporaryDownloadUrlOptions(),
   }) async {
@@ -244,7 +244,7 @@ class S3CompatCloudStorage extends CloudStorage {
 
   @override
   Future<void> deleteFile({
-    required CloudStorageSession session,
+    required Session session,
     required String path,
   }) async {
     final response = await _client.deleteObject(path);
@@ -255,7 +255,7 @@ class S3CompatCloudStorage extends CloudStorage {
 
   @override
   Future<UploadDescription> createUploadDescription({
-    required CloudStorageSession session,
+    required Session session,
     required String path,
     UploadOptions options = const UploadOptions(),
   }) async {
@@ -279,7 +279,7 @@ class S3CompatCloudStorage extends CloudStorage {
 
   @override
   Future<bool> verifyUpload({
-    required CloudStorageSession session,
+    required Session session,
     required String path,
   }) async {
     try {
