+++ integrations/serverpod_cloud_storage_s3_compat/lib/src/upload/multipart_post_strategy.dart
@@ -3,7 +3,7 @@ import 'dart:typed_data';
 import 'package:amazon_cognito_identity_dart_2/sig_v4.dart';
 import 'package:http/http.dart' as http;
 import 'package:path/path.dart' as p;
-import 'package:serverpod_cloud_storage/serverpod_cloud_storage.dart';
+import 'package:serverpod/serverpod.dart';
 
 import '../client/exceptions.dart';
 import '../config/s3_endpoint_config.dart';
