+++ integrations/serverpod_cloud_storage_s3_compat/lib/src/upload/s3_upload_strategy.dart
@@ -1,6 +1,6 @@
 import 'dart:typed_data';
 
-import 'package:serverpod_cloud_storage/serverpod_cloud_storage.dart';
+import 'package:serverpod/serverpod.dart';
 
 import '../client/exceptions.dart';
 import '../config/s3_endpoint_config.dart';
