+++ integrations/serverpod_cloud_storage_s3_compat/pubspec.yaml
@@ -19,7 +19,7 @@ dependencies:
   http: '>=1.1.0 <2.0.0'
   mime: ^2.0.0
   path: ^1.8.3
-  serverpod_cloud_storage: 4.0.0-rc.1
+  serverpod: 4.0.0-rc.1
   serverpod_client: 4.0.0-rc.1
 
 dev_dependencies:
