+++ packages/serverpod/lib/serverpod.dart
@@ -29,7 +29,9 @@ export 'package:serverpod/src/generated/log_level.dart';
 export 'package:serverpod/src/util/migrate_session_logs.dart';
 
 // Cloud storage
-export 'package:serverpod_cloud_storage/serverpod_cloud_storage.dart';
+export 'package:serverpod/src/cloud_storage/cloud_storage.dart';
+export 'package:serverpod/src/cloud_storage/cloud_storage_exception.dart';
+export 'package:serverpod/src/cloud_storage/upload_description.dart';
 export 'package:serverpod/src/cloud_storage/database_cloud_storage.dart';
 
 // Cache
