+++ packages/serverpod/lib/src/server/session.dart
@@ -17,7 +17,7 @@ typedef WillCloseListener = FutureOr<void> Function(Session session);
 /// When a call is made to the [Server] a [Session] object is created. It
 /// contains all data associated with the current connection and provides
 /// easy access to the database.
-abstract class Session implements DatabaseSession, CloudStorageSession {
+abstract class Session implements DatabaseSession {
   final LinkedHashSet<WillCloseListener> _willCloseListeners = LinkedHashSet();
 
   /// Adds a listener that will be called when the session is about to close.
