--- packages/serverpod_cloud_storage/.gitignore
@@ -1,7 +1,0 @@
-# https://dart.dev/tools/private-files
-# Created by `dart pub`
-.dart_tool/
-
-# Avoid committing pubspec.lock for library packages; see
-# https://dart.dev/tools/pub/private-files#pubspec-lock.
-pubspec.lock
