--- packages/serverpod_cloud_storage/README.md
@@ -1,9 +1,0 @@
-![Serverpod banner](https://github.com/serverpod/serverpod/raw/main/misc/images/github-header.webp)
-
-# Serverpod
-This package is a core part of Serverpod. For documentation, visit: [https://docs.serverpod.dev](https://docs.serverpod.dev).
-
-## What is Serverpod?
-Serverpod is an open-source, scalable app server, written in Dart for the Flutter community. Check it out!
-
-[Serverpod.dev](https://serverpod.dev)
