--- packages/serverpod_cloud_storage/analysis_options.yaml
@@ -1,1 +1,0 @@
-include: package:serverpod_lints/public.yaml
