--- packages/serverpod_cloud_storage/pubspec.yaml
@@ -1,18 +1,0 @@
-# This file is generated. Do not modify, instead edit the files in the templates/pubspecs directory.
-# Mode: development
-
-name: serverpod_cloud_storage
-version: 4.0.0-rc.1
-description: Shared cloud storage abstractions for Serverpod integrations.
-repository: https://github.com/serverpod/serverpod
-homepage: https://serverpod.dev
-issue_tracker: https://github.com/serverpod/serverpod/issues
-
-environment:
-  sdk: '^3.12.2'
-
-resolution: workspace
-
-dev_dependencies:
-  serverpod_lints: 4.0.0-rc.1
-  test: ^1.25.5
