+++ templates/pubspecs/integrations/serverpod_cloud_storage_gcp/pubspec.yaml
@@ -20,7 +20,7 @@ dependencies:
   mime: ^2.0.0
   path: ^1.8.3
   pointycastle: ">=3.7.4 <5.0.0"
-  serverpod_cloud_storage: SERVERPOD_VERSION
+  serverpod: SERVERPOD_VERSION
   serverpod_cloud_storage_s3_compat: SERVERPOD_VERSION
 
 dev_dependencies:
