+++ templates/pubspecs/integrations/serverpod_cloud_storage_s3_compat/pubspec.yaml
@@ -18,7 +18,7 @@ dependencies:
   http: '>=1.1.0 <2.0.0'
   mime: ^2.0.0
   path: ^1.8.3
-  serverpod_cloud_storage: SERVERPOD_VERSION
+  serverpod: SERVERPOD_VERSION
   serverpod_client: SERVERPOD_VERSION
 
 dev_dependencies:
