--- templates/pubspecs/packages/serverpod_cloud_storage/pubspec.yaml
@@ -1,17 +1,0 @@
-# TEMPLATE
-
-name: serverpod_cloud_storage
-version: SERVERPOD_VERSION
-description: Shared cloud storage abstractions for Serverpod integrations.
-repository: https://github.com/serverpod/serverpod
-homepage: https://serverpod.dev
-issue_tracker: https://github.com/serverpod/serverpod/issues
-
-environment:
-  sdk: DART_VERSION
-
-resolution: workspace
-
-dev_dependencies:
-  serverpod_lints: SERVERPOD_VERSION
-  test: ^1.25.5
