+++ util/all_packages
@@ -35,7 +35,6 @@ declare -a DART_PACKAGES_PATHS=(
     "packages/serverpod_serialization"
     "packages/serverpod_service_client"
     "packages/serverpod_shared"
-    "packages/serverpod_cloud_storage"
     "packages/serverpod_database"
     "packages/serverpod_embedded_postgres"
     "packages/serverpod_test"
