+++ util/run_tests_analyze
@@ -27,7 +27,6 @@ declare -a projectPaths=(
     "packages/serverpod_serialization"
     "packages/serverpod_service_client"
     "packages/serverpod_shared"
-    "packages/serverpod_cloud_storage"
     "packages/serverpod_database"
     "packages/serverpod_embedded_postgres"
     "packages/serverpod_flutter"
